Revert "feat(auth): passwordless email OTP behind console_auth_redesign (#3178)"
This reverts commit 21b8fe76abe88d7c6a1c849b2b00b0608b54c513.
