+++ apps/deploy-web/src/pages/api/auth/password-signup.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
 
+import { setSession } from "@src/lib/auth0/setSession/setSession";
 import { defineApiHandler } from "@src/lib/nextjs/defineApiHandler/defineApiHandler";
 import { verifyCaptcha } from "@src/middleware/verify-captcha/verify-captcha";
 
@@ -47,7 +48,7 @@ export default defineApiHandler({
     });
 
     if (result.ok) {
-      await services.setSession(req, res, result.val);
+      await setSession(req, res, result.val);
       res.status(204).end();
       return;
     }
