--- apps/deploy-web/tests/ui/passwordless-login.spec.ts
@@ -1,35 +1,0 @@
-import { test } from "./fixture/onboarding-test";
-import { testEnvConfig } from "./fixture/test-env.config";
-import { AuthPagePasswordless } from "./pages/AuthPagePasswordless";
-import { MailsacCodeVerificationStrategy } from "./services/email-verification/mailsac-code.strategy";
-
-test.describe("Passwordless login", () => {
-  let testUserId: string | undefined;
-  const otp = new MailsacCodeVerificationStrategy(testEnvConfig.MAILSAC_API_KEY);
-
-  test.afterEach(async ({ auth0 }) => {
-    if (testUserId) {
-      await auth0.deleteUser(testUserId).catch(() => {});
-      testUserId = undefined;
-    }
-  });
-
-  test("receives OTP, verifies, lands authenticated", async ({ page, auth0 }) => {
-    test.setTimeout(3 * 60 * 1000);
-
-    const email = otp.generateEmail();
-    const auth = new AuthPagePasswordless(page);
-
-    await auth.goto();
-    await auth.startWithEmail(email);
-    await auth.waitForVerifyScreen();
-
-    await otp.verify({ context: page.context(), email, userId: "" });
-
-    await auth.waitForRedirectAwayFromLogin();
-
-    const auth0User = await auth0.getUserByEmail(email);
-    if (!auth0User) throw new Error(`Auth0 user was not created for ${email}`);
-    testUserId = auth0User.user_id;
-  });
-});
